feat(engine-dom): introduce a renderer-factory that can produce a renderer (#2997)
* feat: introduce a renderer-factory that can produce a renderer

* fix: remove htmlPropertyToAttribute usage in @lwc/engine-dom

* chore: inline createFragment into renderer-factory

* feat: move insertStylesheet out of renderer-factory

* fix: move shared tokens out of renderer-factory

* fix: handle base proto inheritance inline in the renderer-factory

* fix: use logWarn instead of console.warn
