+++ packages/@lwc/engine-core/src/framework/main.ts
@@ -48,6 +48,7 @@ export { freezeTemplate } from './freeze-template';
 export { getComponentConstructor } from './get-component-constructor';
 
 // Types -------------------------------------------------------------------------------------------
+export type { RendererAPI } from './renderer';
 export type {
     ConfigValue as WireConfigValue,
     ContextValue as WireContextValue,
