+++ packages/@lwc/engine-dom/src/index.ts
@@ -49,3 +49,4 @@ export { createElement } from './apis/create-element';
 export { isNodeFromTemplate } from './apis/is-node-from-template';
 export { LightningElement } from './apis/lightning-element';
 export { renderer } from './renderer';
+export { rendererFactory } from './renderer-factory';
