+++ packages/@lwc/integration-karma/test/component/unknown-properties/index.spec.js
@@ -7,7 +7,7 @@ describe('unknown properties', () => {
         expect(() => {
             document.body.appendChild(elm);
         }).toLogWarningDev(
-            'Unknown public property "propertyThatDefinitelyDoesNotExist" of element <x-child>. ' +
+            'Error: [LWC warn]: Unknown public property "propertyThatDefinitelyDoesNotExist" of element <x-child>. ' +
                 'This is either a typo on the corresponding attribute "property-that-definitely-does-not-exist", or ' +
                 'the attribute does not exist in this browser or DOM implementation.'
         );
